All list song

+++ backend/app/Http/Controllers/SongController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\MovedSong;
 use App\Models\Song;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
@@ -37,20 +38,57 @@ class SongController extends Controller
         return response()->json(compact('data'));
     }
 
+    public function store_moved(Request $request)
+    {
+        $validator = Validator::make($request->all(), [
+            'nameSong' => 'required|string',
+            'avatarUrl' => 'required|url',
+            'mp3Url' => 'required|url|unique:moved_songs',
+            'describes' => 'required|string',
+            'author' => 'required|string',
+            'views' => 'required|integer',
+            'user_id' => 'required|integer',
+            'singer_id' => 'required|string',
+            'category_id' => 'required|integer',
+            'album_id' => 'required|integer',
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json($validator->errors()->toJson(), 400);
+        }
+
+        $data = new MovedSong();
+        $data->fill($request->all());
+        $data->save();
+        foreach (json_decode($request->singer_id, true) as $s_id) {
+            $data->singers()->attach($s_id);
+        }
+        return response()->json(compact('data'));
+    }
+
     public function index()
     {
         $songs = DB::table('songs')->join('categories', 'songs.category_id', '=', 'categories.id')->select('songs.*', 'songs.category_id')->get();
         return response()->json($songs);
     }
 
-    public function show($id)
+    public function show($user_id, Request $request, UserController $userController)
     {
+        if (!$user_id) {
+            return response()->json(['error' => 'User ID not found.'], 400);
+        }
+        $token = $userController->getAuthenticatedUser();
+        if ($token->getData()->user->id !== (int)$user_id) {
+            $userController->removeToken($request, $request->bearerToken());
+            return response()->json(['error' => 'invalid_access'], 400);
+        }
+
         $songs = DB::table('songs')
             ->select('songs.*', 'users.username', 'categories.category_name', 'albums.album_name')
             ->join('users', 'users.id', '=', 'songs.user_id')
             ->join('albums', 'albums.id', '=', 'songs.album_id')
             ->join('categories', 'categories.id', '=', 'songs.category_id')
-            ->where('users.id', '=', $id)
+            ->where('users.id', '=', (int)$user_id)
             ->get()
             ->toArray();
         $data = [];
@@ -94,14 +132,36 @@ class SongController extends Controller
         return response()->json(compact('data'), 200);
     }
 
+    public function allSongsByID($id)
+    {
+        $res = Song::with('singers')->get()->toArray();
+        $data = array_filter($res, function ($row) use ($id) {
+            return $row['user_id'] === (int)$id;
+        });
+        $data = array_values($data);
+        return response()->json(compact('data'), 200);
+    }
+
+    public function allMovedSongs()
+    {
+        $data = MovedSong::with('singers')->get();
+        return response()->json(compact('data'), 200);
+    }
+
     public function showidsong($id)
     {
         $song = Song::find($id);
         return response()->json($song);
     }
 
-    public function update(Request $request, $id)
+    public function update(Request $request, $id, UserController $userController)
     {
+        $token = $userController->getAuthenticatedUser();
+        if (!$this->getUserIDbySongID($request->id) || ($token->getData()->user->id !== $this->getUserIDbySongID($request->id))) {
+            $userController->removeToken($request, $request->bearerToken());
+            return response()->json(['error' => 'invalid_access'], 400);
+        }
+
         $validator = Validator::make($request->all(), [
             'nameSong' => 'required|string',
             'describes' => 'required|string',
@@ -131,12 +191,56 @@ class SongController extends Controller
         return response()->json(compact('data'));
     }
 
-    public function destroy(Request $request)
+    public function getUserIDbySongID($song_id)
     {
+        $data = DB::table('songs')->where('id', '=', $song_id)->first();
+        return $data->user_id;
+    }
+
+    public function getUserIDbyMovedSongID($song_id)
+    {
+        $data = DB::table('moved_songs')->where('id', '=', $song_id)->first();
+        return $data->user_id;
+    }
+
+    public function destroy(Request $request, UserController $userController)
+    {
+        $token = $userController->getAuthenticatedUser();
+        if (!$request->user_id || ($request->user_id !== $token->getData()->user->id)) {
+            $userController->removeToken($request, $request->bearerToken());
+            return response()->json(['error' => 'User ID invalid or not found.'], 400);
+        }
+        if (!$this->getUserIDbySongID($request->id) ||
+            ($token->getData()->user->id !== $this->getUserIDbySongID($request->id))
+        ) {
+            $userController->removeToken($request, $request->bearerToken());
+            return response()->json(['error' => 'invalid_access'], 400);
+        }
+
         $song = Song::findOrFail($request->id);
         $song->singers()->detach();
         $song->delete();
         return response()->json($song);
+    }
+
+    public function destroyMoved(Request $request, UserController $userController)
+    {
+        $token = $userController->getAuthenticatedUser();
+        if (!$request->user_id || ($request->user_id !== $token->getData()->user->id)) {
+            $userController->removeToken($request, $request->bearerToken());
+            return response()->json(['error' => 'User ID invalid or not found.'], 400);
+        }
+        if (!$this->getUserIDbyMovedSongID($request->id) ||
+            ($token->getData()->user->id !== $this->getUserIDbyMovedSongID($request->id))
+        ) {
+            $userController->removeToken($request, $request->bearerToken());
+            return response()->json(['error' => 'invalid_access'], 400);
+        }
+
+        $movedSong = MovedSong::findOrFail($request->id);
+        $movedSong->singers()->detach();
+        $movedSong->delete();
+        return response()->json($movedSong);
     }
 
     public function search(Request $request)
